server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pillow==11.0.0
numpy==2.2.1
python-multipart==0.0.20
aiofiles==24.1.0
